Expose log.With func
