+++ internal/log/replay_logger.go
@@ -78,5 +78,5 @@ func (l *ReplayLogger) Error(msg string, keyvals ...interface{}) {
 
 // With returns new logger the prepend every log entry with keyvals.
 func (l *ReplayLogger) With(keyvals ...interface{}) log.Logger {
-	return NewReplayLogger(With(l.logger, keyvals), l.isReplay, l.enableLoggingInReplay)
+	return NewReplayLogger(log.With(l.logger, keyvals), l.isReplay, l.enableLoggingInReplay)
 }
