+++ log/with_logger.go
@@ -28,3 +28,45 @@ package log
 type WithLogger interface {
 	With(keyvals ...interface{}) Logger
 }
+
+// With returns Logger instance that prepend every log entry with keyvals. If logger implements WithLogger it is used, otherwise every log call will be intercepted.
+func With(logger Logger, keyvals ...interface{}) Logger {
+	if wl, ok := logger.(WithLogger); ok {
+		return wl.With(keyvals...)
+	}
+
+	return newWithLogger(logger, keyvals...)
+}
+
+type withLogger struct {
+	logger  Logger
+	keyvals []interface{}
+}
+
+func newWithLogger(logger Logger, keyvals ...interface{}) *withLogger {
+	return &withLogger{logger: logger, keyvals: keyvals}
+}
+
+func (l *withLogger) prependKeyvals(keyvals []interface{}) []interface{} {
+	return append(l.keyvals, keyvals...)
+}
+
+// Debug writes message to the log.
+func (l *withLogger) Debug(msg string, keyvals ...interface{}) {
+	l.logger.Debug(msg, l.prependKeyvals(keyvals)...)
+}
+
+// Info writes message to the log.
+func (l *withLogger) Info(msg string, keyvals ...interface{}) {
+	l.logger.Info(msg, l.prependKeyvals(keyvals)...)
+}
+
+// Warn writes message to the log.
+func (l *withLogger) Warn(msg string, keyvals ...interface{}) {
+	l.logger.Warn(msg, l.prependKeyvals(keyvals)...)
+}
+
+// Error writes message to the log.
+func (l *withLogger) Error(msg string, keyvals ...interface{}) {
+	l.logger.Error(msg, l.prependKeyvals(keyvals)...)
+}
